Rename ApiRequest to API Client. Change Client to ZTTP. Fix Response
