+++ example/index.php
@@ -14,24 +14,33 @@ $dotenv->load();
 //    'signature_key' => getenv('RAUDHAH_X_SIGNATURE')
 //])
 //    ->bill()
-//    ->makeBill()
+//    ->makeBill("N09HEFP1")
 //    ->setCustomer("Amirul", "Amirul", "seed.email93@gmail.com", "60123456789", "Melaka")
 //    ->setReference("Testing")
 //    ->setProduct("Product 1", 10.30, 1)
-//    ->create("GU0O6HT7")
-//    ->output();
+//    ->create();
 
-//$response = RaudhahPay::make()
+//$response = RaudhahPay::make([
+//    'api_key' => getenv('RAUDHAH_API_KEY'),
+//    'signature_key' => getenv('RAUDHAH_X_SIGNATURE')
+//])
 //    ->collection()
-//    ->create("Collection Name");
+//    ->create("Tsting123444");
 
 $response = RaudhahPay::make([
     'api_key' => getenv('RAUDHAH_API_KEY'),
     'signature_key' => getenv('RAUDHAH_X_SIGNATURE')
 ])
-    ->directPay()
-    ->payee("GU0O6HT7")
-    ->getDirectPays();
+    ->customer()
+    ->getList();
+
+//$response = RaudhahPay::make([
+//    'api_key' => getenv('RAUDHAH_API_KEY'),
+//    'signature_key' => getenv('RAUDHAH_X_SIGNATURE')
+//])
+//    ->directPay()
+//    ->payee("GU0O6HT7")
+//    ->getDirectPays();
 
 
 header('Content-type: application/json');
