+++ src/RPay/RPCustomer.php
@@ -5,14 +5,14 @@ namespace Afiqiqmal\Rpclient\RPay;
 
 
 use Afiqiqmal\Rpclient\Contracts\RPContracts;
-use Afiqiqmal\Rpclient\HttpClient\ApiRequest;
+use Afiqiqmal\Rpclient\HttpClient\APIClient;
 use Afiqiqmal\Rpclient\HttpClient\PayResponse;
 use Afiqiqmal\Rpclient\Utils\RPUtils;
 
 class RPCustomer implements RPContracts
 {
     /**
-     * @var ApiRequest
+     * @var APIClient
      */
     protected $client;
 
@@ -128,9 +128,9 @@ class RPCustomer implements RPContracts
     }
 
     /**
-     * @return ApiRequest
+     * @return APIClient
      */
-    public function getClient(): ApiRequest
+    public function getClient(): APIClient
     {
         return $this->client;
     }
